Python3 update; Add pipeline configurations to modules; Add postprocessing stages to AR5 modules

+++ modules/ar5/glacierscmip6/ar5_project_glacierscmip6.py
@@ -117,8 +117,8 @@ def ar5_project_glacierscmip6(rng_seed, nmsamps, ntsamps, pipeline_id):
 		zgl=project_glacier1(zit,glparm[igl]['factor'],glparm[igl]['exponent'])
 		
 		# Store these samples
-		ifirst=igl*nrpergl
-		ilast=ifirst+nrpergl
+		ifirst=int(igl*nrpergl)
+		ilast=int(ifirst+nrpergl)
 		glacier[ifirst:ilast,...] = zgl[np.newaxis,:,:] + (mgl * r[ifirst:ilast,np.newaxis] * cvgl)[:,np.newaxis,:]
